refactor: password 입력 기능 비즈니스 로직으로 분리
